Print format DC

+++ ohmgroups/ohm_groups/doctype/dc_not_for_sales/dc_not_for_sales.py
@@ -18,6 +18,27 @@ def on_insert(party):
 	else:
 		return frappe.get_value("Warehouse", {'parent_warehouse': f'Supplier Warehouse - {abbr}', 'warehouse_name':party}, pluck="name")
 
+
+@frappe.whitelist()
+def address_company(company):
+    com_add = frappe.db.get_value("Dynamic Link", {"parenttype":"Address","link_doctype":"Company","link_name":company},"parent")
+    return com_add
+
+@frappe.whitelist()
+def address_shipping(party_type, party):
+    com_add = frappe.db.get_list("Dynamic Link", {"parenttype":"Address","link_doctype":party_type,"link_name":party},pluck="parent")
+    for i in com_add:
+        ship_add =  frappe.get_value('Address',{'address_type':'Shipping','name':i},"name")
+        if ship_add:
+            return ship_add
+@frappe.whitelist()
+def address_billing(party_type, party):
+    com_add = frappe.db.get_list("Dynamic Link", {"parenttype":"Address","link_doctype":party_type,"link_name":party},pluck="parent")
+    for i in com_add:
+        bill_add =  frappe.get_value('Address',{'address_type':'Billing','name':i},"name")
+        if bill_add:
+            return bill_add
+    
 class DCNotforSales(Document):
 
 	
@@ -38,11 +59,13 @@ class DCNotforSales(Document):
 			document.submit()
 
 	def on_cancel(self):
-		frappe.get_doc("Stock Entry",{"dc_no":self.name}).cancel()
+		if frappe.db.exists("Stock Entry",{"dc_no":self.name}):
+			frappe.get_doc("Stock Entry",{"dc_no":self.name}).cancel()
 		
 	def on_trash(self):
-		frappe.get_doc("Stock Entry",{"dc_no":self.name}).delete()
-  
+		if frappe.db.exists("Stock Entry",{"dc_no":self.name}):
+			frappe.get_doc("Stock Entry",{"dc_no":self.name}).delete()
+
 	def validate(self):
 		total_amt = 0
 		total_qty = 0
@@ -51,3 +74,5 @@ class DCNotforSales(Document):
 			total_qty += i.qty 
 		self.total_qty = total_qty
 		self.total_amount = total_amt
+  
+
